Expose a `cleanup` method
Consumers should have a means to unbind handlers bound to global events.
This allows them to avoid memory leaks and prevents the accumulation of
meaningless event handlers for plugin instances that have been
destroyed.

In order to use this API, consumers will need to access the plugin
instance as it is attached to elements with jQuery's "data" API. Because
of this, the attribute name should be a formal committment of this
library and explicitly documented.