Added Pointer for entire range
I think this is helpful particularly for desktop in that clicking on the range can be easier with a mouse that clicking and dragging the handle. Adding the Pointer style for the whole range thus clues the user into the fact that the entire range is clickable.

I also removed the Pointer from the handle class as it seems redundant with having added it to the whole range.