Do not calculate position on hidden elements
If an slider is hidden on init, it will always set its min attribute as its value. This overwrites any default value that is set via the value attribute.

This commit will prevent a hidden element from calculating its position and instead rely on the defaults set. 

This will also prevent any the onSlide and onSlideEnd callbacks from being run on a hidden element - however if it's hidden, these wouldn't be trigger-able anyway!